Removed `.astype('float')` from the final ia.fromarray() call
Gives a different error: python doesn't understand the `pixels` input now

+++ 138_spw25_masking.py
@@ -6,7 +6,6 @@ print("Now masking the cube at sigma threshold")
 # Use the code below to mask the cube at a certain sigma threshold
 cube = SpectralCube.read('source_ab_138_spw25_dirty_512.image')
 mask = cube > 7.0*u.mJy/u.beam
-
 from casatools import image
 ia = image()
 boolmask = mask.include().compute()
@@ -44,5 +43,5 @@ print("Now calculating the eroded/dilated mask")
 # Use the code below to output an eroded/dilated mask
 boolmask_e3_d4 = scipy.ndimage.binary_dilation(scipy.ndimage.binary_erosion(boolmask, iterations=3), iterations=4)
 print("Now outputting the eroded/dilated mask")
-ia.fromarray(outfile='source_ab_138_spw25_dirty_512_e3_d4.mask', pixels=boolmask_e3_d4.astype('float')[:,None,:,:].T, csys=cs.torecord(), overwrite=True)
+ia.fromarray(outfile='source_ab_138_spw25_dirty_512_e3_d4.mask', pixels=boolmask_e3_d4[:,None,:,:].T, csys=cs.torecord(), overwrite=True)
 ia.close()
